Реализовал отправку оповещений-напоминалок пользователям в заданное ими  в уведомлении время. Вынес логику работы с уведомлениями в отдельный сервис NotificationTaskServiceImpl. Добавил новые логи и комментарии.
