+++ src/main/java/pro/sky/telegrambot/constants/Constants.java
@@ -3,4 +3,6 @@ package pro.sky.telegrambot.constants;
 public class Constants {
     public static final String START = "/start";
     public static final String NOTIFICATION = "/notify";
+
+    public static final String NOTIFICATION_PATTERN = "([0-9\\.\\:\\s]{16})(\\s)([\\W0-9]+)";
 }
