+++ src/main/java/pro/sky/telegrambot/listener/TelegramBotUpdatesListener.java
@@ -5,19 +5,11 @@ import com.pengrad.telegrambot.UpdatesListener;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-import pro.sky.telegrambot.model.NotificationTask;
-import pro.sky.telegrambot.repository.NotificationTaskRepository;
+import pro.sky.telegrambot.service.NotificationTaskService;
 
 import javax.annotation.PostConstruct;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
-import java.util.Collection;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import static pro.sky.telegrambot.constants.Constants.*;
 
@@ -25,12 +17,11 @@ import static pro.sky.telegrambot.constants.Constants.*;
 @Slf4j
 public class TelegramBotUpdatesListener implements UpdatesListener {
     private final TelegramBot telegramBot;
+    private final NotificationTaskService notificationTaskService;
 
-    private final NotificationTaskRepository notificationTaskRepository;
-
-    public TelegramBotUpdatesListener(TelegramBot telegramBot, NotificationTaskRepository notificationTaskRepository) {
+    public TelegramBotUpdatesListener(TelegramBot telegramBot, NotificationTaskService notificationTaskService) {
         this.telegramBot = telegramBot;
-        this.notificationTaskRepository = notificationTaskRepository;
+        this.notificationTaskService = notificationTaskService;
     }
 
     @PostConstruct
@@ -48,8 +39,6 @@ public class TelegramBotUpdatesListener implements UpdatesListener {
             Long userChatId = update.message().chat().id();
             //  получаем имя пользователя
             String userName = update.message().chat().firstName();
-            //  метод для определения паттерна уведомления и сохранения его в БД
-            notificationMaker(userChatId, userMessageText);
 
             //  отпраляем ответы пользователю на его команды"
             switch (userMessageText) {
@@ -62,31 +51,17 @@ public class TelegramBotUpdatesListener implements UpdatesListener {
                     log.info("The \"{}\" command was received", userMessageText);
                     sendMessage(userChatId,
                             "Напишите уведомление в формате (чч.мм.гггг чч:мм Текст), "
-                                    + " и мы отправим вам сообщение-напоминалку равно в этот срок");
+                                    + " и мы отправим вам сообщение-напоминалку четко в срок");
                     break;
             }
 
-
+            /*  вызывается метод сервиса уведомлений, который проверит получаемое сообщение на соответствие паттерну
+             *  уведомления и, если соответствие будет выявлено, сохранит в БД.
+             *  Так же производится мониторинг акутальных записей и
+             *  и рассылка уведомлений в нужное время в нужные чаты*/
+            notificationTaskService.notificationMaker(userChatId, userMessageText);
         });
         return UpdatesListener.CONFIRMED_UPDATES_ALL;
-    }
-
-    //  с помощью шедулинга каждую минуту делаем из БД выборку записей, соответствующих настоящему времени
-    @Scheduled(cron = "0 0/1 * * * *")
-    public void run() {
-        log.info("Scheduled method for fetching messages by current time is running");
-        LocalDateTime now = LocalDateTime.now().truncatedTo(ChronoUnit.MINUTES);
-        Collection<NotificationTask> actual = notificationTaskRepository.findAllByDateTimeEquals(now);
-        log.info("The following sample was obtained: {}", actual);
-        sendNotify(actual);
-    }
-
-//  метод проходит по списку всех, вытащенных записей из БД, и высылает уведомления в соответствующие чаты
-    private void sendNotify(Collection<NotificationTask> actual) {
-        for (NotificationTask nt : actual) {
-            sendMessage(nt.getChatId(),
-                    "Внимание! Сработало напоминание! \n\"" + nt.getNotificationMessage() + "\"");
-        }
     }
 
     //  метод для формирования ответа и отправки его пользователю
@@ -95,24 +70,4 @@ public class TelegramBotUpdatesListener implements UpdatesListener {
         telegramBot.execute(sendMessage);
         log.info("The message \"{}\" was sent to chat with id={}", textToSend, chatId);
     }
-
-    //  метод для определения паттерна уведомления и сохранения его в БД
-    private void notificationMaker(Long userChatId, String userMessageText) {
-        //  создаем паттерн для распознавания даты и текста напоминания
-        Pattern pattern = Pattern.compile("([0-9\\.\\:\\s]{16})(\\s)([\\W0-9]+)");
-        //  выполняем поиск совпадений с паттерном
-        Matcher matcher = pattern.matcher(userMessageText);
-        //  если команда соответствует паттерну, вычленяем нужные нам фрагменты, формируем обьект и сохраняем его в БД
-        if (matcher.matches()) {
-            log.info("A command corresponding to the pattern was received. Command: \"{}\"", userMessageText);
-            String dateTimeStr = matcher.group(1);
-            LocalDateTime dateTime = LocalDateTime.parse(dateTimeStr,
-                    DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm"));
-            String notificationMessage = matcher.group(3);
-            notificationTaskRepository.save(new NotificationTask(userChatId, notificationMessage, dateTime));
-            log.info("Notification created and saved to database");
-            sendMessage(userChatId, "Уведомление \"" + userMessageText + "\" создано!");
-        }
-    }
-
 }
