+++ src/main/java/pro/sky/telegrambot/service/NotificationTaskService.java
@@ -1,0 +1,5 @@
+package pro.sky.telegrambot.service;
+
+public interface NotificationTaskService {
+    void notificationMaker(Long userChatId, String userMessageText);
+}
